requirements.txt
click==8.1.3
Faker==15.3.3
Flask==2.2.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.19.0
mysql-connector-python==8.0.31
packaging==21.3
protobuf==3.20.1
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.11.0
